LRCI-1809 Move SpiraCustomListValue into its own class
