+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/spira/SpiraTestCaseRun.java
@@ -352,8 +352,8 @@ public class SpiraTestCaseRun extends BaseSpiraArtifact {
 							getSpiraProject(), SpiraTestCaseRun.class,
 							spiraCustomProperty.getName());
 
-					SpiraCustomList.Value spiraCustomListValue =
-						SpiraCustomList.createSpiraCustomListValue(
+					SpiraCustomListValue spiraCustomListValue =
+						SpiraCustomListValue.createSpiraCustomListValue(
 							getSpiraProject(), spiraCustomList,
 							spiraCustomPropertyValue.getName());
